pyproject.toml
[project]
name = "simple_md"
version = "0.0.1"
description = "Simple Molecular dynamics benchmark"
requires-python = "==3.12.*"
readme = { file = "README.md", content-type = "text/markdown" }

[tool.mypy]
explicit_package_bases = true
disable_error_code = ["import-untyped"]

[[tool.mypy.overrides]]
module = "cloudpathlib.*"
ignore_errors = true
follow_imports = "skip"

[tool.poetry]
name = "example-ml-project"
version = "0.1.0"
description = ""
authors = ["mx-e <mail@mx-e.net>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.4.0"
loguru = "^0.7.2"
wandb = "^0.17.5"
hydra-zen = "^0.13.0"
python-dotenv = "^1.0.1"
submitit = "^1.5.2"
frozendict = "^2.4.6"
tensorflow-datasets = "^4.9.7"
ase = "^3.23.0"
scikit-learn = "^1.5.2"
load-atoms = "^0.3.9"
tensorflow = "^2.18.0"

[tool.poetry.group.datasets.dependencies]
sgdml = "^1.0.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 119
indent-width = 4

[tool.ruff.lint]
ignore = [
    "ANN001",
    "ANN002",
    "ANN003",
    "ANN401",
    "D413",
    "COM812",
    "D100",
    "D103",
    "D104",
    "D105",
    "D107",
    "D205",
    "PD901",
    "D400",
    "D401",
    "D415",
    "FA",
    "SLF",
    "INP",
    "TRY003",
    "TRY201",
    "EM",
    "FBT",
    "RET",
    "C406",
    "E501",
    "E741",
    "PLR2004",
    "RUF009",
    "RUF012",
    "BLE001",
    "N817",
    "N812",
    "S603",
    "S607",
    "S506",
    "FIX002",
    "NPY002",
    "G004",
    "S311",
    "PIE790",
    "TRY400",
    "S108",
    "S101",
    "W191",
    "E111",
    "E114",
    "E117",
    "D101",
    "D102",
    "D206",
    "D300",
    "Q000",
    "Q001",
    "Q002",
    "Q003",
    "COM812",
    "COM819",
    "D203",
    "D213",
    "N806",
    "N803",
    "E712",
    "PLR0913",
    "TC001",
    "PD011",
    "TD002",
    "TD003",
]
select = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"**/tests/**/*.py" = ["S101", "ARG", "FBT"]

[tool.pytest.ini_options]
pythonpath = ["scripts", "scripts/lib"]
